+++ fpdf/fpdf.py
@@ -990,6 +990,23 @@ class FPDF:
             )
         )
 
+    @check_page
+    def circle(self, x, y, r, style=None):
+        """
+        Outputs a circle.
+        It can be drawn (border only), filled (with no border) or both.
+
+        Args:
+            x (int): Abscissa of upper-left bounging box.
+            y (int): Ordinate of upper-left bounging box.
+            r (int): Radius of the circle.
+            style (int): Style of rendering. Possible values are:
+                * `D` or None: draw border. This is the default value.
+                * `F`: fill
+                * `DF` or `FD`: draw and fill
+        """
+        self.ellipse(x, y, r, r, style)
+
     def add_font(self, family, style="", fname=None, uni=False):
         """
         Imports a TrueType, OpenType or Type1 font and makes it available