+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and [PEP 440](https://www.python.org/dev/peps/pep-0440/).
 ## [2.4.2] - not released yet
 ### Added
 - disable font caching when `fpdf.FPDF` constructor invoked with `font_cache_dir=None`
+- `FPDF.circle`: new method added.
 
 ## [2.4.1] - 2021-06-12
 ### Fixed